refactor: batch 모듈에서 필요없는 거 제거
