+++ batch/src/main/java/store/_0982/batch/domain/grouppurchase/GroupPurchaseRepository.java
@@ -3,6 +3,7 @@ package store._0982.batch.domain.grouppurchase;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import store._0982.common.domain.grouppurchase.GroupPurchase;
+import store._0982.common.domain.grouppurchase.GroupPurchaseStatus;
 
 import java.time.OffsetDateTime;
 import java.util.List;
