--- batch/src/main/java/store/_0982/batch/domain/grouppurchase/GroupPurchaseStatus.java
@@ -1,8 +1,0 @@
-package store._0982.batch.domain.grouppurchase;
-
-public enum GroupPurchaseStatus {
-    SCHEDULED,       // 진행전
-    OPEN,            // 진행중
-    SUCCESS,         // 성공
-    FAILED           // 실패
-}
