+++ batch/src/main/java/store/_0982/batch/domain/sellerbalance/SellerBalanceHistoryRepository.java
@@ -1,17 +1,12 @@
 package store._0982.batch.domain.sellerbalance;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import store._0982.common.domain.sellerbalance.SellerBalanceHistory;
 
 import java.util.List;
-import java.util.UUID;
 
 public interface SellerBalanceHistoryRepository {
 
     void save(SellerBalanceHistory sellerBalanceHistory);
-
-    Page<SellerBalanceHistory> findAllMemberId(UUID memberId, Pageable pageable);
 
     void saveAll(List<SellerBalanceHistory> sellerBalanceHistories);
 }
