+++ batch/src/main/java/store/_0982/batch/infrastructure/grouppurchase/GroupPurchaseRepositoryAdaptor.java
@@ -6,8 +6,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 import store._0982.batch.domain.grouppurchase.GroupPurchaseRepository;
-import store._0982.batch.domain.grouppurchase.GroupPurchaseStatus;
 import store._0982.common.domain.grouppurchase.GroupPurchase;
+import store._0982.common.domain.grouppurchase.GroupPurchaseStatus;
 
 import java.time.OffsetDateTime;
 import java.util.List;
