+++ batch/src/main/java/store/_0982/batch/infrastructure/sellerbalance/SellerBalanceHistoryJpaRepository.java
@@ -1,12 +1,9 @@
 package store._0982.batch.infrastructure.sellerbalance;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import store._0982.common.domain.sellerbalance.SellerBalanceHistory;
 
 import java.util.UUID;
 
 public interface SellerBalanceHistoryJpaRepository extends JpaRepository<SellerBalanceHistory, UUID> {
-    Page<SellerBalanceHistory> findAllByMemberId(UUID memberId, Pageable pageable);
 }
