+++ batch/src/main/java/store/_0982/batch/infrastructure/sellerbalance/SellerBalanceHistoryRepositoryAdapter.java
@@ -1,14 +1,11 @@
 package store._0982.batch.infrastructure.sellerbalance;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 import store._0982.batch.domain.sellerbalance.SellerBalanceHistoryRepository;
 import store._0982.common.domain.sellerbalance.SellerBalanceHistory;
 
 import java.util.List;
-import java.util.UUID;
 
 @RequiredArgsConstructor
 @Repository
@@ -19,11 +16,6 @@ public class SellerBalanceHistoryRepositoryAdapter implements SellerBalanceHisto
     @Override
     public void save(SellerBalanceHistory sellerBalanceHistory) {
         sellerBalanceHistoryJpaRepository.save(sellerBalanceHistory);
-    }
-
-    @Override
-    public Page<SellerBalanceHistory> findAllMemberId(UUID memberId, Pageable pageable) {
-        return sellerBalanceHistoryJpaRepository.findAllByMemberId(memberId, pageable);
     }
 
     @Override
